[Issue 624] Execution / Tet suite page still don't refresh when running an execution
Replaced most of href="#" by href="javascript:void(0)" so that the execution /
test suite page url does not change.

Another solution would be for each javascript handler to return false so that
the href="#" is not evaluated. Which is the weaker, previous approach.

Also fixed : "Start" menu looses its skin when an item is removed
from test plan